requirements.txt
SQLAlchemy==1.4.41
numpy==1.23.3
pandas==1.4.4
requests==2.26.0
pendulum==2.1.2
apache-airflow[amazon]==2.3.4
clickhouse-driver==0.2.4
clickhouse-sqlalchemy==0.2.2
airflow-clickhouse-plugin==0.10.0
dbt-clickhouse==1.4.0
dbt-core==1.4.4
airflow-dbt==0.4.0
cachelib==0.9.0
urllib3==1.26.16